Make L_obj for stages (L component)

The following parameters will be set to the values in options. If they are not found, default values will be used.

\(\psi_E\) or psi_E: egg maturation (hatching) rate

\(\phi_E\) or phi_E: egg mortality rate

\(\psi_i\) or psi1, psi2, psi3, psi4: larval instar maturation rates

\(\xi_i\) or xi1, xi2, xi3, xi4: delayed maturation responses to crowding

\(\phi_i\) or phi1, phi2, phi3, phi4: density-independent larval mortality rates

\(\theta_i\) or theta1, theta2, theta3, theta4: density-dependent larval mortality slopes

\(\phi_P\) or phi_P: pupal mortality rate

\(\psi_P\) or psi_P: pupal maturation (emergence) rate

\(w_1, w_2, w_3\) or w1, w2, w3: instar crowding weights

make_L_obj_stages(
  nHabitats,
  options = list(),
  psi_E = 1/2,
  phi_E = 1/10,
  psi1 = 1/2,
  xi1 = 0,
  phi1 = 1/2,
  theta1 = 1/100,
  psi2 = 1/2,
  xi2 = 0,
  phi2 = 1/2,
  theta2 = 1/100,
  psi3 = 1/2,
  xi3 = 0,
  phi3 = 1/2,
  theta3 = 1/100,
  psi4 = 1/2,
  xi4 = 0,
  phi4 = 1/2,
  theta4 = 1/100,
  phi_P = 1/2,
  psi_P = 1/3,
  w1 = 0.3,
  w2 = 0.5,
  w3 = 0.8
)

Arguments

nHabitats

the number of habitats in the model

options

a named list

psi_E

egg maturation (hatching) rate

phi_E

egg mortality rate

psi1

first instar maturation rate

xi1

first instar delayed maturation response to crowding

phi1

first instar density-independent mortality rate

theta1

first instar density-dependent mortality slope

psi2

second instar maturation rate

xi2

second instar delayed maturation response to crowding

phi2

second instar density-independent mortality rate

theta2

second instar density-dependent mortality slope

psi3

third instar maturation rate

xi3

third instar delayed maturation response to crowding

phi3

third instar density-independent mortality rate

theta3

third instar density-dependent mortality slope

psi4

fourth instar maturation rate

xi4

fourth instar delayed maturation response to crowding

phi4

fourth instar density-independent mortality rate

theta4

fourth instar density-dependent mortality slope

phi_P

pupal mortality rate

psi_P

pupal maturation (emergence) rate

w1

contribution of \(L_1\) to mean crowding

w2

contribution of \(L_2\) to mean crowding

w3

contribution of \(L_3\) to mean crowding

Value

L_obj an L component object

See also